Array JS names
Added the feature to grab array names inside <script> tags. very common to see parameter names there that later got reflected or can be modified from GET requests. this helped me to find a 10k XSS on VzM. And is something that Param Miner already does automatically would be great to add this to Arjun and this few lines does.